CM6: Move markdown-related files into a markdown directory
