+++ .gitignore
@@ -467,20 +467,20 @@ packages/app-mobile/utils/fs-driver-rn.js
 packages/app-mobile/utils/setupNotifications.js
 packages/app-mobile/utils/shareHandler.js
 packages/app-mobile/utils/types.js
-packages/editor/CodeMirror/computeSelectionFormatting.js
 packages/editor/CodeMirror/configFromSettings.js
 packages/editor/CodeMirror/createEditor.test.js
 packages/editor/CodeMirror/createEditor.js
-packages/editor/CodeMirror/decoratorExtension.js
-packages/editor/CodeMirror/markdownCommands.bulletedVsChecklist.test.js
-packages/editor/CodeMirror/markdownCommands.test.js
-packages/editor/CodeMirror/markdownCommands.toggleList.test.js
-packages/editor/CodeMirror/markdownCommands.js
-packages/editor/CodeMirror/markdownMathParser.test.js
-packages/editor/CodeMirror/markdownMathParser.js
-packages/editor/CodeMirror/markdownReformatter.test.js
-packages/editor/CodeMirror/markdownReformatter.js
-packages/editor/CodeMirror/syntaxHighlightingLanguages.js
+packages/editor/CodeMirror/markdown/computeSelectionFormatting.js
+packages/editor/CodeMirror/markdown/decoratorExtension.js
+packages/editor/CodeMirror/markdown/markdownCommands.bulletedVsChecklist.test.js
+packages/editor/CodeMirror/markdown/markdownCommands.test.js
+packages/editor/CodeMirror/markdown/markdownCommands.toggleList.test.js
+packages/editor/CodeMirror/markdown/markdownCommands.js
+packages/editor/CodeMirror/markdown/markdownMathParser.test.js
+packages/editor/CodeMirror/markdown/markdownMathParser.js
+packages/editor/CodeMirror/markdown/markdownReformatter.test.js
+packages/editor/CodeMirror/markdown/markdownReformatter.js
+packages/editor/CodeMirror/markdown/syntaxHighlightingLanguages.js
 packages/editor/CodeMirror/testUtil/createEditor.js
 packages/editor/CodeMirror/testUtil/forceFullParse.js
 packages/editor/CodeMirror/testUtil/loadLanguages.js
