+++ packages/editor/CodeMirror/markdown/syntaxHighlightingLanguages.ts
