+++ packages/editor/CodeMirror/theme.ts
@@ -8,7 +8,7 @@ import { tags } from '@lezer/highlight';
 import { EditorView } from '@codemirror/view';
 import { Extension } from '@codemirror/state';
 
-import { inlineMathTag, mathTag } from './markdownMathParser';
+import { inlineMathTag, mathTag } from './markdown/markdownMathParser';
 
 // For an example on how to customize the theme, see:
 //
